requirements.txt
flask
pillow
numpy
tensorflow
python-dotenv
flask-cors